ai-services/requirements.txt
# AI Services Requirements
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
redis==5.0.1
numpy==1.26.2
scikit-learn==1.3.2
pillow==10.1.0
easyocr==1.7.0
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
sentence-transformers==2.2.2
python-dotenv==1.0.0
gunicorn==21.2.0
huggingface-hub==0.20.3
